# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HighTac Local Platform API
  version: 2.0.0
  summary: Contract for the HighTac LAN management platform.
  description: |
    The Windows backend is the source of truth for stations, tags, products,
    bindings, commands, devices, and audit records. Timestamps are RFC 3339 UTC.

    Admin browser requests use the `hightac_session` cookie. Every mutating
    admin request also requires `X-CSRF-Token`. Approved Android installations
    use bearer identity and never receive MQTT credentials. The canonical retry
    key for retry-safe writes is the `Idempotency-Key` header; keys are scoped to
    the authenticated principal and retained for at least 24 hours.

    Approved Android devices also use bearer-authenticated, allowlisted mobile
    proxies for OpenAI-compatible Responses and JianDaoYun catalog reads. Third-
    party API credentials remain server-side and are never returned to clients.
  contact:
    name: HighTac Engineering
  license:
    name: Proprietary
...
servers:
  - url: http://{host}:{port}/api/v1
    description: HighTac platform on the trusted local network
    variables:
      host:
        default: 192.168.1.105
      port:
        default: "8088"
...
paths:
...
  /mobile/media/{token}:
    get:
      tags: [Mobile proxy]
      operationId: getMobileMedia
      summary: Fetch a short-lived signed JianDaoYun image through HighTac.
      description: |
        The path token is HMAC-signed with a key derived from the server's
        device-enrollment secret and expires after a short interval. This GET
        intentionally does not require an Android Authorization header so image
        loaders can use it directly. The server accepts only HTTPS media URLs
        on the exact `files.jiandaoyun.com` host, does not follow redirects, and
        returns only bounded `image/*` responses.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: Opaque, short-lived signed media capability.
          schema:
            type: string
            minLength: 16
            maxLength: 8192
      responses:
        "200":
          description: Validated image bytes from the allowlisted media host.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            Cache-Control:
              schema:
                type: string
                const: private, no-store
            X-Content-Type-Options:
              schema:
                type: string
                const: nosniff
            Content-Security-Policy:
              schema:
                type: string
                const: sandbox; default-src 'none'
          content:
            "image/*":
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
  /backups:
    get:
      tags: [Backups]
      operationId: listBackups
      summary: List platform-created database backups.
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/BackupStatus"
      responses:
        "200":
          description: Paginated backup records.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [Backups]
      operationId: createBackup
      summary: Start an immediate SQLite online backup.
      security:
        - AdminSession: []
          CsrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Backup started.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "507":
          $ref: "#/components/responses/InsufficientStorage"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: hightac_session
      description: Opaque HttpOnly SameSite=Strict administrator session cookie.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required together with AdminSession on every mutating browser request.
    AndroidBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-device-token
      description: Approved installation token stored in Android Keystore; only its hash is stored server-side.
    EnrollmentPollSecret:
      type: apiKey
      in: header
      name: X-Enrollment-Secret
      description: Short-lived installation-scoped secret returned once when enrollment is created.
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: One-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: Number of records per page.
      schema:
        type: integer
        enum: [20, 50, 100]
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Comma-separated allowlisted fields; prefix a field with `-` for descending order.
      schema:
        type: string
        maxLength: 256
        example: -updated_at
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID. Reusing a key with the same method, path, actor,
        and normalized body returns the original result. Reusing it with a
        different request returns 409 IDEMPOTENCY_KEY_REUSED.
      schema:
        type: string
        format: uuid
      example: 2ec5fd45-0e4d-4be4-a6bd-3517118b64df
...
  headers:
    RequestId:
      description: Correlation identifier for logs and support diagnostics.
      schema:
        type: string
        format: uuid
    Location:
      description: Relative API path of the created or accepted resource.
      schema:
        type: string
        example: /api/v1/light-commands/d7d74ed4-44d7-4934-af3a-a80d93aaf63f
  responses:
...
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Valid authentication is required.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Forbidden:
      description: The authenticated actor lacks permission, CSRF validation failed, or the Android device is not approved.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: FORBIDDEN
              message: This actor cannot perform the requested operation.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    NotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Conflict:
      description: Current resource state conflicts with the request or an idempotency key was reused for a different request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            activeBinding:
              value:
                error:
                  code: TAG_ALREADY_BOUND
                  message: The tag already has an active binding; explicit rebind is required.
                  details:
                    - field: tag_id
                      code: ACTIVE_BINDING_EXISTS
                      message: Tag AD100000048F is bound to product 1711A-OLD.
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            idempotencyMismatch:
              value:
                error:
                  code: IDEMPOTENCY_KEY_REUSED
                  message: The idempotency key belongs to a different request.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            migrationPreviewMismatch:
              value:
                error:
                  code: MIGRATION_PREVIEW_MISMATCH
                  message: The preview token does not match this device and record set.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
            migrationPreviewStale:
              value:
                error:
                  code: MIGRATION_PREVIEW_STALE
                  message: Authoritative binding state changed after the preview.
                  details: []
                  request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    Gone:
      description: An enrollment, one-time delivery window, or signed media link has expired.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PayloadTooLarge:
      description: A request, uploaded workbook, or proxied media response exceeds its configured size limit.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnsupportedMediaType:
      description: Request content or a proxied media response has an unsupported media type.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationError:
      description: One or more path, query, header, or body values failed validation.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: VALIDATION_ERROR
              message: One or more values are invalid.
              details:
                - field: tag_id
                  code: PATTERN_MISMATCH
                  message: Must match ^AD1[0-9A-F]{9}$.
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
...
    ServiceUnavailable:
      description: A required platform component such as the MQTT bridge or Windows service controller is unavailable.
      headers:
        Retry-After:
          description: Suggested retry delay in seconds.
          schema:
            type: integer
            minimum: 1
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: MQTT_UNAVAILABLE
              message: The broker is not available; no command was created.
              details: []
              request_id: b6d35a55-63ac-4b0e-99ae-78262b27c7da
    BadGateway:
      description: The configured third-party service failed or rejected its server credential.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    GatewayTimeout:
      description: The configured third-party service did not respond before the server timeout.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InsufficientStorage:
      description: The platform data volume lacks space for the backup or pre-restore copy.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Uuid:
      type: string
      format: uuid
      example: 2b16ae29-f7ab-4c44-9fc6-a0ca693d9c68
    Timestamp:
      type: string
      format: date-time
      example: "2026-07-16T08:15:30Z"
...
    Sha256:
      type: string
      pattern: ^[0-9a-f]{64}$
      example: 9e107d9d372bb6826bd81d3542a419d6e6299f6ae24c42b2762c9b6f55d25e5f
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      additionalProperties: false
      required: [code, message, details, request_id]
      properties:
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]{1,63}$
        message:
          type: string
          minLength: 1
          maxLength: 1024
        details:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          $ref: "#/components/schemas/Uuid"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: [string, "null"]
          maxLength: 256
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]{1,63}$
        message:
          type: string
          minLength: 1
          maxLength: 1024
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          enum: [20, 50, 100]
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    BackupStatus:
      type: string
      enum: [RUNNING, SUCCEEDED, FAILED, DELETED]
    Backup:
      type: object
      additionalProperties: false
      required: [id, filename, reason, status, size_bytes, sha256, schema_version, created_at, completed_at, error_code]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        filename:
          type: string
          minLength: 1
          maxLength: 255
          description: Basename only; no arbitrary filesystem path is exposed.
        reason:
          type: string
          enum: [SCHEDULED, MANUAL, PRE_RESTORE]
        status:
          $ref: "#/components/schemas/BackupStatus"
        size_bytes:
          type: [integer, "null"]
          minimum: 0
        sha256:
          oneOf:
            - $ref: "#/components/schemas/Sha256"
            - type: "null"
        schema_version:
          type: [string, "null"]
          maxLength: 64
        created_at:
          $ref: "#/components/schemas/Timestamp"
        completed_at:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        error_code:
          type: [string, "null"]
          maxLength: 64
    BackupPage:
      type: object
      additionalProperties: false
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Backup"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...